requirements.txt
ultralytics==8.1.0    # YOLOv11核心库
numpy==1.26.0         # 数值计算，轨迹点坐标计算
pandas==2.2.0         # 数据处理和轨迹数据存储
matplotlib==3.8.2     # 数据可视化
opencv-python==4.8.0  # 视频读取、绘图、目标框可视化
shapely==2.0.2        # 多边形区域 (ROI) 判断
cvzone==1.5.6         # OpenCV 辅助工具，便于绘制轨迹及交互
pyyaml==6.0.1         # 配置文件 (如 `botsort.yaml`) 的读取

